pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "annqc"
version = "0.1.0"
description = "Reproducible QC reports for AnnData single-cell RNA-seq workflows"
requires-python = ">=3.9"
dependencies = [
    "scanpy>=1.9",
    "anndata>=0.9",
    "scrublet>=0.2",
    "click>=8.0",
    "pyyaml>=6.0",
    "jinja2>=3.0",
    "plotly>=5.0",
]

[project.scripts]
annqc = "annqc.cli:main"

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "black", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["annqc*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
